feat: update users while login
